Refactor depcheck.py to improve package update handling and add changelog writing functionality

+++ debian/deploy.sh
@@ -11,17 +11,8 @@ if [[ $RELEASE_TYPE != "testing" && $RELEASE_TYPE != "stable" ]]; then
     exit 1
 fi
 
-PACKAGE=$2
-if [[ -z $PACKAGE ]] || [[ ! -d "$SCRIPT_DIR/$PACKAGE" ]]; then
-    echo "Error: Package not provided or not found in directory: $SCRIPT_DIR/$PACKAGE"
-    echo "Usage: deploy.sh [testing|stable] <package>"
-    exit 1
-fi
-
-set -eu
-
 if [[ $RELEASE_TYPE == "stable" ]]; then
-    read -p "Are you sure you want to deploy a PRODUCTION release? (y/n) " -n 1 -r
+    read -p "Are you sure you want to deploy to PRODUCTION? (y/n) " -r
     echo
     if [[ ! $REPLY =~ ^[Yy]$ ]]
     then
@@ -29,22 +20,34 @@ if [[ $RELEASE_TYPE == "stable" ]]; then
     fi
 fi
 
-rm -rf "$REPO_ROOT/build/debian/${PACKAGE}_"*.deb
+PACKAGES=$2
+IFS=',' read -ra PACKAGE_ARRAY <<< "$PACKAGES"
+for PACKAGE in "${PACKAGE_ARRAY[@]}"; do
+    if [[ -z $PACKAGE ]] || [[ ! -d "$SCRIPT_DIR/$PACKAGE" ]]; then
+        echo "Error: Package not provided or not found in directory: $SCRIPT_DIR/$PACKAGE"
+        echo "Usage: deploy.sh [testing|stable] <package>"
+        exit 1
+    fi
 
-# Build the package
-bash $SCRIPT_DIR/build.sh $PACKAGE
+    set -eu
 
-cd $DEB_REPO_ROOT
+    rm -rf "$REPO_ROOT/build/debian/${PACKAGE}_"*.deb
 
-rm -rf "$DEB_REPO_ROOT/docs/pool/$RELEASE_TYPE/main/${PACKAGE}_"*.deb 2>/dev/null || true
-cp "$REPO_ROOT/build/debian/${PACKAGE}_"*.deb "$DEB_REPO_ROOT/docs/pool/$RELEASE_TYPE/main/"
+    # Build the package
+    bash $SCRIPT_DIR/build.sh $PACKAGE
+
+    cd $DEB_REPO_ROOT
+
+    rm -rf "$DEB_REPO_ROOT/docs/pool/$RELEASE_TYPE/main/${PACKAGE}_"*.deb 2>/dev/null || true
+    cp "$REPO_ROOT/build/debian/${PACKAGE}_"*.deb "$DEB_REPO_ROOT/docs/pool/$RELEASE_TYPE/main/"
+done
 
 # Recompile the repository with the updated/new package
 ./build.sh $RELEASE_TYPE
 
 git add --all
 git status
-read -p "Would you like to push these changes? (y/n) " -n 1 -r
+read -p "Would you like to push these changes? (y/n) " -r
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
